python-in-tauri/resources/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
llama-cpp-python>=0.2.25
appdirs>=1.4.4
requests>=2.31.0